requirements.txt
fastapi
uvicorn
google-generativeai
requests
beautifulsoup4
graphviz
pydantic
python-dotenv
googlesearch-python